Remove YAML heads for the cloud-config bits
Looks like RancherOS uses the old form of cloud-config multiple stanza
handling where the last instance of the stanza is the one that has
precedence. The "rancher" section follows this format, at least.

I'm removing the YAML headings, while keeping the formatting, this will
allow us to render the "write_files" and the "rancher" sections with any
configured services or rendered Dockerfiles consolidated in
terraform_rancher_user_data, in addition to any specific runtime
configuration we need for rancher.