Add network_mode option
This will allow one to set net=host, basically.

Also, not using a heredoc for the YAML output anymore for the services
section. Unfortunately, HIL does not provide a way currently to trim
newlines, and having conditionals on newlines breaks up the YAML. For
now doing to work around this with a personal vim macro and leave the
squashed string in there. Put in a feature request to get this looked at
in HIL (may take a stab at it myself).